Make config.router defined by default
